Feat/be/mate

+++ server/bf/src/main/java/com/bobfriends/bf/comment/service/CommentService.java
@@ -1,11 +1,13 @@
 package com.bobfriends.bf.comment.service;
 
 
+import com.bobfriends.bf.auth.jwt.JwtTokenizer;
 import com.bobfriends.bf.comment.dto.CommentDto;
 import com.bobfriends.bf.comment.entity.Comment;
 import com.bobfriends.bf.comment.repository.CommentRepository;
 import com.bobfriends.bf.exception.BusinessLogicException;
 import com.bobfriends.bf.exception.ExceptionCode;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
@@ -13,12 +15,12 @@ import java.util.Objects;
 import java.util.Optional;
 
 @Service
+@RequiredArgsConstructor
 public class CommentService {
     private final CommentRepository commentRepository;
+    private final JwtTokenizer jwtTokenizer;
 
-    public CommentService(CommentRepository commentRepository){
-        this.commentRepository = commentRepository;
-    }
+
 
     public Comment createComment(Comment comment){
 
@@ -46,10 +48,22 @@ public class CommentService {
         return findVerifiedComment(commentId);
     }
 
-    public void deleteComment(Long commentId){
+    public void deleteComment(Long commentId,String token){
         // 존재하는 댓글인지 검증
-        findVerifiedComment(commentId);
-        commentRepository.deleteById(commentId);
+        Comment findComment = findVerifiedComment(commentId);
+
+        String base64EncodedSecretKey = jwtTokenizer.encodeBase64SecretKey(jwtTokenizer.getSecretKey());
+
+        long loginId = jwtTokenizer.getMemberIdFromToken(token, base64EncodedSecretKey);
+
+        long writerMemberId = findComment.getMember().getMemberId();
+
+        // 로그인 한 회원이 작성자이면
+        if(loginId == writerMemberId){
+            commentRepository.deleteById(commentId);
+        }else {
+            throw new RuntimeException("등록한 작성자가 아닙니다");
+        }
     }
 
     private Comment findVerifiedComment(Long commentId) {
